+++ src/main/java/me/ramidzkh/mekae2/integration/jei/AMJEIPlugin.java
@@ -1,7 +1,8 @@
-package me.ramidzkh.mekae2;
+package me.ramidzkh.mekae2.integration.jei;
 
 import net.minecraft.resources.ResourceLocation;
 
+import me.ramidzkh.mekae2.AppliedMekanistics;
 import me.ramidzkh.mekae2.ae2.ChemicalIngredientConverter;
 import mekanism.api.IMekanismAccess;
 import mezz.jei.api.IModPlugin;
