--- src/main/resources/appmek.mixins.json
@@ -1,10 +1,0 @@
-{
-  "minVersion": "0.8",
-  "required": true,
-  "package": "me.ramidzkh.mekae2.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
